starting with modifying local testing env to have datetimes in correct format
