+++ v6_LinReg_py/__init__.py
@@ -40,21 +40,22 @@ def RPC_fit_round(db_client, coefs, intercepts, data_cols, extra_cols, lr, seed,
     try:
         connection = psycopg2.connect(PG_URI)
         cursor = connection.cursor()
+        info("connected to PG database. building dataframe..")
 
         #incrementally build dataframe
         for col in all_cols:
             data_pg = cursor.execute("SELECT {} FROM ncdc".format(col))
-            data_df[col] = data_pg.fetchall()
+            data_df[col] = cursor.fetchall()
     except psycopg2.Error as e:
         return e
 
-    info("connected to PG database")
-
+    info("dataframe built")
     data = data_df[all_cols].dropna()
 
     blood_dates = data["date_blood"].values
     mri_dates = data["date_mri"].values
 
+    info("blood dates: " + str(blood_dates[0]))
     blood_dates_years = np.array([int(blood_date[:4]) for blood_date in blood_dates])
     mri_dates_years = np.array([int(mri_date[:4]) for mri_date in mri_dates])
 
@@ -127,3 +128,110 @@ def RPC_fit_round(db_client, coefs, intercepts, data_cols, extra_cols, lr, seed,
 
 
 
+### I probably won't need these functions anymore, but just in case I decided to keep them in this file as an archive
+
+import os
+from vantage6.tools.util import info
+import psycopg2
+
+
+def RPC_pgdb_print(db_client, PG_URI = None):
+
+        
+    # could also use this to set URI
+    if PG_URI == None:
+        PG_URI = 'postgresql://{}:{}@{}:{}'.format(
+            os.getenv("PGUSER"),
+            os.getenv("PGPASSWORD"),
+            os.getenv("PGHOST"),
+            os.getenv("PGPORT"))
+
+
+   
+    info("connecting to PG DB:" + str(PG_URI))
+
+    # get data from postgres DB
+    try:
+        connection = psycopg2.connect(PG_URI)
+        cursor = connection.cursor()
+
+        info("connected to PG database")
+        data_pg = cursor.execute("SELECT * FROM ncdc")
+        info(str(cursor.fetchall()))
+        data = cursor.fetchall()
+        cursor.close()
+        connection.close()
+        
+        return data
+    except psycopg2.Error as e:
+        info("error: " + str(e))
+
+        cursor.close()
+        connection.close()
+        return e
+
+
+
+
+
+
+def RPC_create_db(db_client, PG_URI = None):
+
+
+    rng = np.random.default_rng()
+    if PG_URI == None:
+        PG_URI = 'postgresql://{}:{}@{}:{}'.format(
+            os.getenv("PGUSER"),
+            os.getenv("PGPASSWORD"),
+            os.getenv("PGHOST"),
+            os.getenv("PGPORT"))
+
+    info("connecting to PG DB:" + str(PG_URI))
+
+    try:
+        connection = psycopg2.connect(PG_URI)
+        #print(connection)
+
+        cursor = connection.cursor()
+
+        info("connected")
+        all_cols =  ["metabo_age", "brain_age", "date_blood", "date_mri", "birth_year", "sex", "dm", "bmi", "education_category"]
+        vals = abs(rng.standard_normal(size = (len(all_cols), 30)))
+        #print(len(all_cols))
+
+        
+        info("adding ncdc table")
+        cursor.execute("DROP TABLE IF EXISTS ncdc")
+        cursor.execute("CREATE TABLE ncdc ()")
+        for i, col in enumerate(all_cols):
+            if col == "date_blood" or col == "date_mri":
+                #modify the value in 'vals' for these columns
+                base_time = np.datetime64('2015-01-01')
+                vals[i,:] = base_time + rng.standard_normal(size = len(vals[i,:]))
+            
+            query = "ALTER TABLE ncdc ADD " + col + " int;"
+            cursor.execute(query)
+            #cursor.execute("""ALTER TABLE ncdc ADD %s float;""", (col,))
+        info("adding dummy data to ncdc table")
+        for value in vals.T:
+            #query = "INSERT INTO ncdc '" + col + "' VALUES " + str(value)
+            #cursor.execute(query)
+
+            cursor.execute("INSERT INTO ncdc VALUES {}".format(tuple(value)))
+        
+
+
+        cursor.execute("SELECT * FROM ncdc")
+        data = cursor.fetchall()
+
+        info("done!")
+
+        connection.commit()
+        cursor.close()
+        connection.close()
+
+        return data
+    except psycopg2.Error as e:
+        info("error: " + str(e))
+
+        return e
